[IMRPOVEMENT] Merged ThemesActivity and CustomThemesActivity.
