--- app/src/main/res/layout/activity_create_themes.xml
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="chat.rocket.android.thememanager.ui.CreateThemesActivity">
-
-    <include
-        android:id="@+id/layout_app_bar"
-        layout="@layout/app_bar"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/select_theme"
-        style="@style/ChangeTheme.Content.TextView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:paddingStart="@dimen/theme_screen_edge_left_and_right_padding"
-        android:paddingTop="@dimen/theme_screen_edge_top_padding"
-        android:text="Tap a theme name to apply."
-        app:layout_constraintStart_toStartOf="@+id/layout_app_bar"
-        app:layout_constraintTop_toBottomOf="@+id/layout_app_bar" />
-
-    <TextView
-        android:id="@+id/edit_theme"
-        style="@style/ChangeTheme.Content.TextView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:paddingStart="@dimen/theme_screen_edge_left_and_right_padding"
-        android:text="Tap a palette color to edit."
-        app:layout_constraintStart_toStartOf="@+id/layout_app_bar"
-        app:layout_constraintTop_toBottomOf="@+id/select_theme" />
-
-    <TextView
-        android:id="@+id/delete_theme"
-        style="@style/ChangeTheme.Content.TextView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:paddingStart="@dimen/theme_screen_edge_left_and_right_padding"
-        android:text="Swipe to remove palette."
-        app:layout_constraintStart_toStartOf="@+id/layout_app_bar"
-        app:layout_constraintTop_toBottomOf="@+id/edit_theme" />
-
-    <TextView
-        android:id="@+id/text_custom_themes"
-        style="@style/ChangeTheme.TextView.Headline"
-        android:paddingStart="@dimen/theme_screen_edge_left_and_right_padding"
-        android:paddingTop="@dimen/theme_screen_edge_top_padding"
-        android:text="@string/custom_themes"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/delete_theme"
-        />
-
-    <ScrollView
-        android:id="@+id/scroll_view"
-        android:layout_marginTop="20dp"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:fillViewport="true"
-        app:layout_constraintBottom_toTopOf="@id/layout_item_add_theme"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/text_custom_themes">
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/custom_themes_recycler_view"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        tools:itemCount="2"
-        tools:listitem="@layout/item_theme_row" />
-
-    </ScrollView>
-
-    <include
-        android:id="@+id/layout_item_add_theme"
-        layout="@layout/item_add_theme"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/scroll_view" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
