+++ app/src/main/res/layout/item_theme_info.xml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <TextView
+        android:id="@+id/current_theme"
+        style="@style/ChangeTheme.Content.TextView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:paddingStart="@dimen/theme_screen_edge_left_and_right_padding"
+        android:paddingTop="@dimen/theme_screen_edge_top_padding"
+        android:text="Current Theme"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:id="@+id/divider"
+        style="@style/ChangeTheme.Content.TextView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:paddingTop="@dimen/theme_screen_edge_top_padding"
+        android:text="   |   "
+        app:layout_constraintLeft_toRightOf="@+id/current_theme"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:id="@+id/theme_last_changed"
+        style="@style/ChangeTheme.Content.TextView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:paddingTop="@dimen/theme_screen_edge_top_padding"
+        android:text="@string/last_changed"
+        app:layout_constraintLeft_toRightOf="@+id/divider"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <include
+        android:id="@+id/item_theme_row_small"
+        layout="@layout/item_theme_row_small"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/current_theme" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
