+++ app/src/main/res/layout/item_theme_row_small.xml
@@ -18,17 +18,6 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <ImageView
-        android:id="@+id/item_small_theme_preview_primary_text"
-        android:layout_width="45sp"
-        android:layout_height="5dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="@dimen/screen_edge_left_and_right_margins"
-        android:src="@drawable/theme_preview_small"
-        android:tint="?colorPrimaryText"
-        app:layout_constraintLeft_toRightOf="@+id/item_small_theme_preview_accent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageView
         android:id="@+id/item_small_theme_preview_primary"
         android:layout_width="45sp"
         android:layout_height="5dp"
@@ -36,7 +25,29 @@
         android:layout_marginTop="@dimen/screen_edge_left_and_right_margins"
         android:src="@drawable/theme_preview_small"
         android:tint="?colorPrimary"
-        app:layout_constraintLeft_toRightOf="@+id/item_small_theme_preview_primary_text"
+        app:layout_constraintLeft_toRightOf="@+id/item_small_theme_preview_accent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ImageView
+        android:id="@+id/item_small_theme_preview_primary_dark"
+        android:layout_width="45sp"
+        android:layout_height="5dp"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="@dimen/screen_edge_left_and_right_margins"
+        android:src="@drawable/theme_preview_small"
+        android:tint="?colorPrimaryDark"
+        app:layout_constraintLeft_toRightOf="@+id/item_small_theme_preview_primary"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ImageView
+        android:id="@+id/item_small_theme_preview_primary_text"
+        android:layout_width="45sp"
+        android:layout_height="5dp"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="@dimen/screen_edge_left_and_right_margins"
+        android:src="@drawable/theme_preview_small"
+        android:tint="?colorPrimaryText"
+        app:layout_constraintLeft_toRightOf="@+id/item_small_theme_preview_primary_dark"
         app:layout_constraintTop_toTopOf="parent" />
 
     <ImageView
@@ -47,18 +58,7 @@
         android:layout_marginTop="@dimen/screen_edge_left_and_right_margins"
         android:src="@drawable/theme_preview_small"
         android:tint="?colorDescriptiveText"
-        app:layout_constraintLeft_toRightOf="@+id/item_small_theme_preview_primary"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageView
-        android:id="@+id/item_small_theme_preview_background"
-        android:layout_width="45sp"
-        android:layout_height="5dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="@dimen/screen_edge_left_and_right_margins"
-        android:src="@drawable/theme_preview_small"
-        android:tint="?android:colorBackground"
-        app:layout_constraintLeft_toRightOf="@+id/item_small_theme_preview_descriptive"
+        app:layout_constraintLeft_toRightOf="@+id/item_small_theme_preview_primary_text"
         app:layout_constraintTop_toTopOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
