+++ app/src/main/res/values/dimens.xml
@@ -51,6 +51,7 @@
     <!-- Theme -->
     <dimen name="theme_screen_edge_left_and_right_padding">20dp</dimen>
     <dimen name="theme_screen_edge_top_padding">25dp</dimen>
+    <dimen name="theme_title_top_padding">20dp</dimen>
 
 
 </resources>
